solana_vote_program

Module vote_state

Vote state, vote program Receive and processes votes from validators

Modules

serde_compact_vote_state_update

vote_state_1_14_11

vote_state_versions

Structs

BlockTimestamp

CircBuf

LandedVote

Lockout

Vote

VoteAuthorizeCheckedWithSeedArgs

VoteAuthorizeWithSeedArgs

VoteInit

VoteState

VoteState1_14_11

VoteStateUpdate

Enums

VoteAuthorize

VoteStateVersions

VoteTransaction

Constants

INITIAL_LOCKOUT

MAX_EPOCH_CREDITS_HISTORY

MAX_LOCKOUT_HISTORY

VOTE_CREDITS_GRACE_SLOTS

VOTE_CREDITS_MAXIMUM_PER_SLOT

VOTE_CREDITS_MAXIMUM_PER_SLOT_OLD

Functions

authorize

Authorize the given pubkey to withdraw or sign votes. This may be called multiple times, but will implicitly withdraw authorization from the previously authorized key

create_account

create_account_with_authorized

do_process_vote_state_update

from

initialize_account

Initialize the vote_state for a vote account Assumes that the account is being init as part of a account creation or balance transfer and that the transaction must be signed by the staker’s keys

is_commission_increase

Given a proposed new commission, returns true if this would be a commission increase, false otherwise

is_commission_update_allowed

Given the current slot and epoch schedule, determine if a commission change is allowed

process_new_vote_state

process_slot_vote_unchecked

process_vote

process_vote_state_update

process_vote_unchecked

“unchecked” functions used by tests and Tower

process_vote_unfiltered

process_vote_with_account

to

update_commission

Update the vote account’s commission

update_validator_identity

Update the node_pubkey, requires signature of the authorized voter

withdraw

Withdraw funds from the vote account